fix: fix broken tests and make test parse a hard error
Also several cleanups to the parsing/errors to properly produce spans/context.

I keep checking in broken test files and not noticing because the test harness considers it ok
for a test to fail to parse/type (emitting a warning that gets buried by test output immediately).
there's literally no reason to be permissive in this case, as the tests loading/parsing isn't
platform specific, and there's no reason to toss non-tests in these dirs.